Workaround a pedantic lintian warning in overrides
The comment on the lintian override about the spelling error in the
manpage, was tripping a pedantic warning about spelling in the override
file (talk about meta). Someone else bumped into this and this is
tracked as #968416, but in the meantime just replace that comment to
avoid the extra (pedantic) warning.