Add upstream patches to fix compat with Net::DNS 1.36
Net::DNS is notoriously bad with API stability, which broke gdnsd's test
suite with the last working version being 1.32. Add 3 upstream patches
(soon to be forwarded upstream), one to restore functionality with 1.33
and two for 1.35. Also add a fourth patch to help with an EOF race
identified while testing this.

Closes: #1002225