autopkgtest: mark daemon as flaky; common-ops as skippable
In the Ubuntu autopkgtest environment, systemd-resolved is installed and
bound to port 53, preventing gdnsd from binding to [::]:53, and thus
from starting. This may be the case in other environments as well, or in
Debian in the future.

Mark the "daemon" test as "flaky", and the common-operations test as
"skippable", with a check to skip the test if the daemon is not running.

LP: #1764327